2021-12-07 08:57 UTC+0100 Aleksander Czajczynski (hb fki.pl)   * cimgui/hbarrays.cpp   * ChangeLog.txt     ! correct hb_igIntsPush() c&p typos
  * sokol/hbsokol.c
    + expose sapp_quit(), sapp_request_quit(), sapp_cancel_quit()
      to .prg code

    * handle SAPP_EVENTTYPE_QUIT_REQUESTED event, add ImQuit()
      procedure to your .prg code to get notified about it
      On desktop platform you can call sapp_cancel_quit() from
      the ImQuit() to prevent exiting. For example if
      [x] close button was clicked and you wan to display
      some confirmation popup. This can be also used to hide
      ImGui window instead of closing if your application is
      multithreaded, where another window can be Harbour GT too.

   + examples/procdbf.prg
   + examples/test.dbf
     + added multi-threading example of DO WHILE ! EoF() ; DBSkip()
       file processing using codeblock and statics to interact
       with the display

+++ cimgui/hbarrays.cpp
@@ -205,7 +205,7 @@ HB_FUNC( HB_IGINTSPUSH )
 {
    PHB_IG_INTS pInts = hb_ig_ints_par( 1 );
    int * values = ( pInts ? pInts->pBuf : NULL );
-   int value = ( float ) hb_parni( 2 );
+   int value = hb_parni( 2 );
 
    if( values )
    {
